+++ CHANGES.md
@@ -16,6 +16,8 @@
    [#417](https://github.com/CCI-Tools/cate/issues/417)
  * Values persisted in "Download Data Source" window
    [#307](https://github.com/CCI-Tools/cate/issues/207)
+* Only allow valid python identifiers as resource names
+  [#436](https://github.com/CCI-Tools/cate/issues/436)
 
 ### Other Changes